Move readline prompt to separate func/file
