+++ utils/data/recreate.js
@@ -1,12 +1,9 @@
-const readline = require('readline').createInterface({
-  input: process.stdin,
-  output: process.stdout,
-});
-
 const db = require('./db.js');
 const tables = require('./tables.js');
+const {prompt} = require('./prompt.js');
 
 const con = db.connect();
+const msg = `Are you sure you want to recreate your database? You will lose all your data. y/N  `;
 
 // Fully drops and re-creates the database using the tables from `tables.js`
 async function recreateDB() {
@@ -24,18 +21,4 @@ async function recreateDB() {
   }
 }
 
-readline.question(
-  `Are you sure you want to recreate your database? You will lose all your data. y/N  `,
-  async (answer) => {
-    const lowerAns = answer.toLowerCase();
-    if (lowerAns === 'y' || lowerAns === 'yes') {
-      await recreateDB();
-      readline.close();
-      db.disconnect(con, 'success');
-    } else {
-      console.log('Exiting. Data will not be lost.');
-      readline.close();
-      db.disconnect(con);
-    }
-  }
-);
+prompt(msg, recreateDB, con);
